Created database manipulation class; Solved concurrent db locking issue; Moved Job query & update functions to lib.dbmgr; Connected Job execution system to attack.db & removed associated hardcoded components.
